refactor attendance log to access module
